Made all tests in Users_Service_Tests, made all assertions in test class, fix Content Type Header Test
